마이보드페이지 일부 수정
